I have changed the minimun number of samples to built a model from 10 to 3
#38

Also added a warning if the model used had less than 10 samples and if the oc estimated is below or abouve the initial oc range to estimate the mode. Second and third task in issue #38.
Seeing the warning is dificult because the warning for mangrove data with less than 5% of om prints for every sample insted of ones.
I will try to fix this or add it to the issue tasks.